news-generator/pyproject.toml
[project]
name = "news-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.10.0",
    "langchain~=1.0",
    "langchain-community~=0.4",
    "langchain-openai~=1.0",
    "langgraph~=1.0",
    "python-dotenv>=1.2.1",
]
